@v4-nikishin 2022-12-04
van@Vasiliys-MacBook-Pro-2 hw10_program_optimization % benchstat -delta-test=none optimized origin
name             old time/op    new time/op       delta
GetDomainStat-8    4.75µs ± 0%  137383.73µs ± 0%  +2889147.80%

name             old alloc/op   new alloc/op      delta
GetDomainStat-8    5.93kB ± 0%  134445.11kB ± 0%  +2267867.38%

name             old allocs/op  new allocs/op     delta
GetDomainStat-8      41.0 ± 0%    1700090.0 ± 0%  +4146460.98%